aplicando cadastrar, atualizar, pesquisar e deletar na primeira entidade

+++ src/main/java/br/com/fiap/DAO/EventosDao.java
@@ -2,6 +2,6 @@ package br.com.fiap.DAO;
 
 import br.com.fiap.Entity.Eventos;
 
-public interface EventosDao extends GenericDao<Eventos, Long> {
+public interface EventosDao extends GenericDao<Eventos, Integer> {
 
 }
