+++ src/main/java/br/com/fiap/DAO/EventosDaoImpl.java
@@ -4,7 +4,7 @@ import javax.persistence.EntityManager;
 
 import br.com.fiap.Entity.Eventos;
 
-public class EventosDaoImpl extends GenericDaoImpl<Eventos, Long> implements EventosDao {
+public class EventosDaoImpl extends GenericDaoImpl<Eventos, Integer> implements EventosDao {
 
 	public EventosDaoImpl(EntityManager ent) {
 		super(ent);
