+++ src/main/java/br/com/fiap/DAO/IngressosDao.java
@@ -2,6 +2,6 @@ package br.com.fiap.DAO;
 
 import br.com.fiap.Entity.Ingressos;
 
-public interface IngressosDao extends GenericDao<Ingressos, Long> {
+public interface IngressosDao extends GenericDao<Ingressos, Integer> {
 
 }
