+++ src/main/java/br/com/fiap/DAO/IngressosDaoImpl.java
@@ -4,7 +4,7 @@ import javax.persistence.EntityManager;
 
 import br.com.fiap.Entity.Ingressos;
 
-public class IngressosDaoImpl extends GenericDaoImpl<Ingressos, Long> implements IngressosDao {
+public class IngressosDaoImpl extends GenericDaoImpl<Ingressos, Integer> implements IngressosDao {
 
 	public IngressosDaoImpl(EntityManager ent) {
 		super(ent);
