+++ src/main/java/br/com/fiap/DAO/LocaisDao.java
@@ -2,6 +2,6 @@ package br.com.fiap.DAO;
 
 import br.com.fiap.Entity.Locais;
 
-public interface LocaisDao extends GenericDao<Locais, Long>{
+public interface LocaisDao extends GenericDao<Locais, Integer>{
 
 }
