+++ src/main/java/br/com/fiap/DAO/LocaisDaoImpl.java
@@ -4,7 +4,7 @@ import javax.persistence.EntityManager;
 
 import br.com.fiap.Entity.Locais;
 
-public class LocaisDaoImpl extends GenericDaoImpl<Locais, Long> implements LocaisDao{
+public class LocaisDaoImpl extends GenericDaoImpl<Locais, Integer> implements LocaisDao{
 
 	public LocaisDaoImpl(EntityManager ent) {
 		super(ent);
