+++ src/main/java/br/com/fiap/DAO/OrganizadoresDao.java
@@ -2,6 +2,6 @@ package br.com.fiap.DAO;
 
 import br.com.fiap.Entity.Organizadores;
 
-public interface OrganizadoresDao extends GenericDao<Organizadores, Long> {
+public interface OrganizadoresDao extends GenericDao<Organizadores, Integer> {
 
 }
