+++ src/main/java/br/com/fiap/DAO/ParticipantesDao.java
@@ -2,7 +2,7 @@ package br.com.fiap.DAO;
 
 import br.com.fiap.Entity.Participantes;
 
-public interface ParticipantesDao extends GenericDao<Participantes, Long> {
+public interface ParticipantesDao extends GenericDao<Participantes, Integer> {
 
 
 }
