+++ src/main/java/br/com/fiap/DAO/ParticipantesDaoImpl.java
@@ -4,7 +4,7 @@ import javax.persistence.EntityManager;
 
 import br.com.fiap.Entity.Participantes;
 
-public class ParticipantesDaoImpl extends GenericDaoImpl<Participantes, Long> implements ParticipantesDao {
+public class ParticipantesDaoImpl extends GenericDaoImpl<Participantes, Integer> implements ParticipantesDao {
 
 	public ParticipantesDaoImpl(EntityManager ent) {
 		super(ent);
